refactor code

+++ src/core/utils/smooth.h
@@ -4,8 +4,8 @@
 
 #define smooth_float(current, target, speed)                                   \
   current = ((float)current < (float)target)                                   \
-                ? fmin((float)current + (float)speed, (float)target)           \
-                : fmax((float)current - (float)speed, (float)target);
+                ? fminf((float)current + (float)speed, (float)target)          \
+                : fmaxf((float)current - (float)speed, (float)target);
 
 void smooth_vec(Vector2 *position, Vector2 target, float move);
 void smooth_rect(Rectangle *box, Rectangle target, float move);
